# Conflicts:
#	Truffles/.idea/workspace.xml
#	Truffles/src/OI.cpp
#	Truffles/src/PIDStuff.cpp
#	Truffles/src/RobotMap.cpp
#	Truffles/src/RobotMap.h

+++ Truffles/src/Subsystems/Chassis.h
@@ -1,4 +1,3 @@
-
 #ifndef CHASSIS_H
 #define CHASSIS_H
 
@@ -12,64 +11,62 @@
  *
  * @author ExampleAuthor
  */
-class Chassis : public Subsystem {
- private:
-  std::shared_ptr<CANTalon> leftFront;
-  std::shared_ptr<CANTalon> leftRear;
-  std::shared_ptr<CANTalon> rightFront;
-  std::shared_ptr<CANTalon> rightRear;
-  std::shared_ptr<RobotDrive> drive;
+class Chassis: public Subsystem {
+private:
+	std::shared_ptr<CANTalon> leftFront;
+	std::shared_ptr<CANTalon> leftRear;
+	std::shared_ptr<CANTalon> rightFront;
+	std::shared_ptr<CANTalon> rightRear;
+	std::shared_ptr<RobotDrive> drive;
 
- public:
-  Chassis();
-  void InitDefaultCommand();
-  /**
-   * delegates to MecanumDrive_Cartesian
-   * @param stickPosition
-   * @param gyroAngle
-   */
-  void MecanumDrive(std::shared_ptr<Joystick> stickPosition,
-                    double gyroAngle = 0.0);
-  void AutoDrive(float fwdSpeed, float rotateSpeed);
-  void MecanumDrive_Cartesian(double x,
-                              double y,
-                              double rotation,
-                              double gyroAngle = 0.0);
-  void MecanumDrive_Polar(double magnitude, double direction, double rotation);
-  void Drive(double outputMagnitude, double curve);
-  void TankDrive(double leftValue,
-                 double rightValue,
-                 bool squaredInputs = true);
-  void stop();
+public:
+	Chassis();
+	void InitDefaultCommand();
+	/**
+	 * delegates to MecanumDrive_Cartesian
+	 * @param stickPosition
+	 * @param gyroAngle
+	 */
+	void MecanumDrive(std::shared_ptr<Joystick> stickPosition,
+			double gyroAngle = 0.0);
+	void AutoDrive(double fwdSpeed, double rotateSpeed);
+	void MecanumDrive_Cartesian(double x, double y, double rotation,
+			double gyroAngle = 0.0);
+	void MecanumDrive_Polar(double magnitude, double direction,
+			double rotation);
+	void Drive(double outputMagnitude, double curve);
+	void TankDrive(double leftValue, double rightValue,
+	bool squaredInputs = true);
+	void stop();
 
-  double getRightFrontCurrent();
-  double getRightRearCurrent();
-  double getLeftFrontCurrent();
-  double getLeftRearCurrent();
+	double getRightFrontCurrent();
+	double getRightRearCurrent();
+	double getLeftFrontCurrent();
+	double getLeftRearCurrent();
 
-  double getLeftRearVelocity();
-  double getLeftRearPosition();
-  double getRightRearPosition();
-  double getRightRearVelocity();
-  double getLeftFrontVelocity();
-  double getLeftFrontPosition();
-  double getRightFrontPosition();
-  double getRightFrontVelocity();
+	double getLeftRearVelocity();
+	double getLeftRearPosition();
+	double getRightRearPosition();
+	double getRightRearVelocity();
+	double getLeftFrontVelocity();
+	double getLeftFrontPosition();
+	double getRightFrontPosition();
+	double getRightFrontVelocity();
 
-  double getHeading();
+	double getHeading();
 
-  void zeroEncoders();
+	void zeroEncoders();
 
-  //  CANTalon* getTalons();
+	//  CANTalon* getTalons();
 
-  void toggleRobotFrontDirection();
+	void toggleRobotFrontDirection();
 
-  void prepareForTeleop();
+	void prepareForTeleop();
 
-  //  void initMagicMode();
-  //  void initRegularMode();
-  void dashboardTelemetry();
-  double rotateAngle = 0.0;
+	//  void initMagicMode();
+	//  void initRegularMode();
+	void dashboardTelemetry();
+	double rotateAngle = 0.0;
 };
 
 #endif
